+++ packages/gatsby-plugin-sitemap/src/internals.js
@@ -2,10 +2,11 @@ import fs from "fs"
 import pify from "pify"
 import minimatch from "minimatch"
 
-const withoutTrailingSlash = path =>
+export const withoutTrailingSlash = path =>
   path === `/` ? path : path.replace(/\/$/, ``)
 
 export const writeFile = pify(fs.writeFile)
+export const renameFile = pify(fs.rename)
 
 export const runQuery = (handler, query, excludes, pathPrefix) =>
   handler(query).then(r => {
@@ -27,6 +28,17 @@ export const runQuery = (handler, query, excludes, pathPrefix) =>
       page.node.path = (pathPrefix + page.node.path).replace(/^\/\//g, `/`)
       return page
     })
+
+    if (!r.data.site.siteMetadata.siteUrl) {
+      throw new Error(
+        `SiteMetaData 'siteUrl' property is required. Check out the documentation to see a working example: https://www.gatsbyjs.org/packages/gatsby-plugin-sitemap/#how-to-use`
+      )
+    }
+
+    // remove trailing slash of siteUrl
+    r.data.site.siteMetadata.siteUrl = withoutTrailingSlash(
+      r.data.site.siteMetadata.siteUrl
+    )
 
     return r.data
   })