+++ packages/gatsby-source-wordpress/src/__tests__/http-exception-handler.js
@@ -1,0 +1,7 @@
+const httpExceptionHandler = require(`../http-exception-handler`)
+
+describe(`http-exception-handler`, () => {
+  it(`handles errors that lack responses without crashing`, () => {
+    expect(() => httpExceptionHandler({})).not.toThrowError()
+  })
+})