+++ packages/gatsby/src/db/loki/nodes-query.js
@@ -1,7 +1,6 @@
 const _ = require(`lodash`)
 const prepareRegex = require(`../../utils/prepare-regex`)
 const { getNodeTypeCollection } = require(`./nodes`)
-const sift = require(`sift`)
 const { emitter } = require(`../../redux`)
 
 // Cleared on DELETE_CACHE
@@ -13,49 +12,6 @@ emitter.on(`DELETE_CACHE`, () => {
     delete fieldUsages[field]
   }
 })
-
-// Takes a raw graphql filter and converts it into a mongo-like args
-// object that can be understood by the `sift` library. E.g `eq`
-// becomes `$eq`
-function siftifyArgs(object) {
-  const newObject = {}
-  _.each(object, (v, k) => {
-    if (_.isPlainObject(v)) {
-      if (k === `elemMatch`) {
-        k = `$elemMatch`
-      }
-      newObject[k] = siftifyArgs(v)
-    } else {
-      // Compile regex first.
-      if (k === `regex`) {
-        newObject[`$regex`] = prepareRegex(v)
-      } else if (k === `glob`) {
-        const Minimatch = require(`minimatch`).Minimatch
-        const mm = new Minimatch(v)
-        newObject[`$regex`] = mm.makeRe()
-      } else {
-        newObject[`$${k}`] = v
-      }
-    }
-  })
-  return newObject
-}
-
-// filter nodes using the `sift` library. But isn't this a loki query
-// file? Yes, but we need to support all functionality provided by
-// `run-sift`, and there are some operators that loki can't
-// support. Like `elemMatch`, so for those fields, we fall back to
-// sift
-function runSift(nodes, query) {
-  if (nodes) {
-    const siftQuery = {
-      $elemMatch: siftifyArgs(query),
-    }
-    return sift(siftQuery, nodes)
-  } else {
-    return null
-  }
-}
 
 // Takes a raw graphql filter and converts it into a mongo-like args
 // object that can be understood by loki. E.g `eq` becomes
@@ -96,24 +52,36 @@ function toMongoArgs(gqlFilter, lastFieldType) {
   _.each(gqlFilter, (v, k) => {
     if (_.isPlainObject(v)) {
       if (k === `elemMatch`) {
-        // loki doesn't support elemMatch, so use sift (see runSift
-        // comment above)
-        mongoArgs[`$where`] = obj => {
-          const result = runSift(obj, v)
-          return result && result.length > 0
-        }
+        const gqlFieldType = lastFieldType.ofType
+        mongoArgs[`$elemMatch`] = toMongoArgs(v, gqlFieldType)
       } else {
         const gqlFieldType = lastFieldType.getFields()[k].type
         mongoArgs[k] = toMongoArgs(v, gqlFieldType)
       }
     } else {
-      // Compile regex first.
       if (k === `regex`) {
-        mongoArgs[`$regex`] = prepareRegex(v)
+        const re = prepareRegex(v)
+        // To ensure that false is returned if a field doesn't
+        // exist. E.g `{nested.field: {$regex: /.*/}}`
+        mongoArgs[`$where`] = obj => !_.isUndefined(obj) && re.test(obj)
       } else if (k === `glob`) {
         const Minimatch = require(`minimatch`).Minimatch
         const mm = new Minimatch(v)
         mongoArgs[`$regex`] = mm.makeRe()
+      } else if (k === `eq` && v === null) {
+        mongoArgs[`$in`] = [null, undefined]
+      } else if (
+        k === `eq` &&
+        lastFieldType &&
+        lastFieldType.constructor.name === `GraphQLList`
+      ) {
+        mongoArgs[`$contains`] = v
+      } else if (
+        k === `ne` &&
+        lastFieldType &&
+        lastFieldType.constructor.name === `GraphQLList`
+      ) {
+        mongoArgs[`$containsNone`] = v
       } else if (
         k === `in` &&
         lastFieldType &&
@@ -122,13 +90,14 @@ function toMongoArgs(gqlFilter, lastFieldType) {
         mongoArgs[`$containsAny`] = v
       } else if (
         k === `nin` &&
+        lastFieldType &&
         lastFieldType.constructor.name === `GraphQLList`
       ) {
         mongoArgs[`$containsNone`] = v
       } else if (k === `ne` && v === null) {
         mongoArgs[`$ne`] = undefined
       } else if (k === `nin` && lastFieldType.name === `Boolean`) {
-        mongoArgs[`$nin`] = v.concat([false])
+        mongoArgs[`$nin`] = v.concat([undefined])
       } else {
         mongoArgs[`$${k}`] = v
       }
@@ -168,16 +137,19 @@ function toMongoArgs(gqlFilter, lastFieldType) {
 //     $regex: // as above
 //   }
 // }
-function dotNestedFields(acc, o, path = ``) {
-  if (_.isPlainObject(o)) {
-    if (_.isPlainObject(_.sample(o))) {
-      _.forEach(o, (v, k) => {
-        dotNestedFields(acc, v, path + `.` + k)
-      })
+const toDottedFields = (filter, acc = {}, path = []) => {
+  Object.keys(filter).forEach(key => {
+    const value = filter[key]
+    const nextValue = _.isPlainObject(value) && value[Object.keys(value)[0]]
+    if (key === `$elemMatch`) {
+      acc[path.join(`.`)] = { [`$elemMatch`]: toDottedFields(value) }
+    } else if (_.isPlainObject(nextValue)) {
+      toDottedFields(value, acc, path.concat(key))
     } else {
-      acc[_.trimStart(path, `.`)] = o
+      acc[path.concat(key).join(`.`)] = value
     }
-  }
+  })
+  return acc
 }
 
 // The query language that Gatsby has used since day 1 is `sift`. Both
@@ -187,60 +159,50 @@ function dotNestedFields(acc, o, path = ``) {
 // doesn't exist, is null, or bar is null. Whereas loki will return
 // false if the foo field doesn't exist or is null. This ensures that
 // loki queries behave like sift
-function fixNeTrue(flattenedFields) {
-  return _.transform(flattenedFields, (result, v, k) => {
-    if (v[`$ne`] === true) {
-      const s = k.split(`.`)
-      if (s.length > 1) {
-        result[s[0]] = {
-          $or: [
-            {
-              $exists: false,
-            },
-            {
-              $where: obj => obj === null || obj[s[1]] !== true,
-            },
-          ],
-        }
-        return result
-      }
-    }
-    result[k] = v
-    return result
-  })
+const isNeTrue = (obj, path) => {
+  if (path.length) {
+    const [first, ...rest] = path
+    return obj == null || obj[first] == null || isNeTrue(obj[first], rest)
+  } else {
+    return obj !== true
+  }
 }
 
+const fixNeTrue = filter =>
+  Object.keys(filter).reduce((acc, key) => {
+    const value = filter[key]
+    if (value[`$ne`] === true) {
+      const [first, ...path] = key.split(`.`)
+      acc[first] = { [`$where`]: obj => isNeTrue(obj, path) }
+    } else {
+      acc[key] = value
+    }
+    return acc
+  }, {})
+
 // Converts graphQL args to a loki filter
-function convertArgs(gqlArgs, gqlType) {
-  const dottedFields = {}
-  dotNestedFields(dottedFields, toMongoArgs(gqlArgs.filter, gqlType))
-  return fixNeTrue(dottedFields)
-}
+const convertArgs = (gqlArgs, gqlType) =>
+  fixNeTrue(toDottedFields(toMongoArgs(gqlArgs.filter, gqlType)))
 
 // Converts graphql Sort args into the form expected by loki, which is
 // a vector where the first value is a field name, and the second is a
-// boolean `isDesc`. Nested fields delimited by `___` are replaced by
-// periods. E.g
+// boolean `isDesc`. E.g
 //
 // {
 //   fields: [ `frontmatter___date`, `id` ],
-//   order: `desc`
+//   order: [`desc`]
 // }
 //
 // would return
 //
 // [ [ `frontmatter.date`, true ], [ `id`, false ] ]
 //
-// Note that the GraphQL Sort API provided by Gatsby doesn't allow the
-// order to be specified per field. The sift implementation uses
-// lodash `orderBy`, but only applies the sort order to the first
-// field. So we do the same here
 function toSortFields(sortArgs) {
   const { fields, order } = sortArgs
   const lokiSortFields = []
   for (let i = 0; i < fields.length; i++) {
-    const dottedField = fields[i].replace(/___/g, `.`)
-    const isDesc = i === 0 ? _.lowerCase(order) === `desc` : false
+    const dottedField = fields[i]
+    const isDesc = order[i] && order[i].toLowerCase() === `desc`
     lokiSortFields.push([dottedField, isDesc])
   }
   return lokiSortFields
@@ -269,7 +231,7 @@ function ensureFieldIndexes(coll, lokiArgs) {
  *
  * @param {Object} args. Object with:
  *
- * {Object} gqlType: built during `./build-node-types.js`
+ * {Object} gqlType: A GraphQL type
  *
  * {Object} queryArgs: The raw graphql query as a js object. E.g `{
  * filter: { fields { slug: { eq: "/somepath" } } } }`
@@ -282,7 +244,7 @@ function ensureFieldIndexes(coll, lokiArgs) {
  * @returns {promise} A promise that will eventually be resolved with
  * a collection of matching objects (even if `firstOnly` is true)
  */
-async function runQuery({ gqlType, queryArgs, context = {}, firstOnly }) {
+async function runQuery({ gqlType, queryArgs, firstOnly }) {
   // Clone args as for some reason graphql-js removes the constructor
   // from nested objects which breaks a check in sift.js.
   const gqlArgs = JSON.parse(JSON.stringify(queryArgs))
@@ -291,8 +253,8 @@ async function runQuery({ gqlType, queryArgs, context = {}, firstOnly }) {
   ensureFieldIndexes(coll, lokiArgs)
   let chain = coll.chain().find(lokiArgs, firstOnly)
 
-  if (gqlArgs.sort) {
-    const sortFields = toSortFields(gqlArgs.sort)
+  if (queryArgs.sort) {
+    const sortFields = toSortFields(queryArgs.sort)
 
     // Create an index for each sort field. Indexing requires sorting
     // so we lose nothing by ensuring an index is added for each sort