+++ packages/gatsby/src/utils/api-runner-node.js
@@ -1,6 +1,6 @@
 const Promise = require(`bluebird`)
-const glob = require(`glob`)
 const _ = require(`lodash`)
+const chalk = require(`chalk`)
 
 const tracer = require(`opentracing`).globalTracer()
 const reporter = require(`gatsby-cli/lib/reporter`)
@@ -8,6 +8,15 @@ const getCache = require(`./get-cache`)
 const apiList = require(`./api-node-docs`)
 const createNodeId = require(`./create-node-id`)
 const createContentDigest = require(`./create-content-digest`)
+const {
+  buildObjectType,
+  buildUnionType,
+  buildInterfaceType,
+  buildInputObjectType,
+} = require(`../schema/types/type-builders`)
+const { emitter } = require(`../redux`)
+const { getNonGatsbyCodeFrame } = require(`./stack-trace-utils`)
+const { trackBuildError, decorateEvent } = require(`gatsby-telemetry`)
 
 // Bind action creators per plugin so we can auto-add
 // metadata to actions they create.
@@ -94,12 +103,58 @@ const runAPI = (plugin, api, args) => {
 
     const cache = getCache(plugin.name)
 
+    // Ideally this would be more abstracted and applied to more situations, but right now
+    // this can be potentially breaking so targeting `createPages` API and `createPage` action
+    let actions = doubleBoundActionCreators
+    let apiFinished = false
+    if (api === `createPages`) {
+      let alreadyDisplayed = false
+      const createPageAction = actions.createPage
+      // create new actions object with wrapped createPage action
+      // doubleBoundActionCreators is memoized, so we can't just
+      // reassign createPage field as this would cause this extra logic
+      // to be used in subsequent APIs and we only want to target this `createPages` call.
+      actions = {
+        ...actions,
+        createPage: (...args) => {
+          createPageAction(...args)
+          if (apiFinished && !alreadyDisplayed) {
+            const warning = [
+              reporter.stripIndent(`
+              Action ${chalk.bold(
+                `createPage`
+              )} was called outside of its expected asynchronous lifecycle ${chalk.bold(
+                `createPages`
+              )} in ${chalk.bold(plugin.name)}.
+              Ensure that you return a Promise from ${chalk.bold(
+                `createPages`
+              )} and are awaiting any asynchronous method invocations (like ${chalk.bold(
+                `graphql`
+              )} or http requests).
+              For more info and debugging tips: see ${chalk.bold(
+                `https://gatsby.dev/sync-actions`
+              )}
+            `),
+            ]
+
+            const possiblyCodeFrame = getNonGatsbyCodeFrame()
+            if (possiblyCodeFrame) {
+              warning.push(possiblyCodeFrame)
+            }
+
+            reporter.warn(warning.join(`\n\n`))
+            alreadyDisplayed = true
+          }
+        },
+      }
+    }
+
     const apiCallArgs = [
       {
         ...args,
         pathPrefix,
-        boundActionCreators: doubleBoundActionCreators,
-        actions: doubleBoundActionCreators,
+        boundActionCreators: actions,
+        actions,
         loadNodeContent,
         store,
         emitter,
@@ -114,6 +169,12 @@ const runAPI = (plugin, api, args) => {
         createNodeId: namespacedCreateNodeId,
         createContentDigest,
         tracing,
+        schema: {
+          buildObjectType,
+          buildUnionType,
+          buildInterfaceType,
+          buildInputObjectType,
+        },
       },
       plugin.pluginOptions,
     ]
@@ -125,21 +186,31 @@ const runAPI = (plugin, api, args) => {
         const cb = (err, val) => {
           pluginSpan.finish()
           callback(err, val)
+          apiFinished = true
         }
-        gatsbyNode[api](...apiCallArgs, cb)
+
+        try {
+          gatsbyNode[api](...apiCallArgs, cb)
+        } catch (e) {
+          trackBuildError(api, {
+            error: e,
+            pluginName: `${plugin.name}@${plugin.version}`,
+          })
+          throw e
+        }
       })
     } else {
       const result = gatsbyNode[api](...apiCallArgs)
       pluginSpan.finish()
-      return Promise.resolve(result)
+      return Promise.resolve(result).then(res => {
+        apiFinished = true
+        return res
+      })
     }
   }
 
   return null
 }
-
-let filteredPlugins
-const hasAPIFile = plugin => glob.sync(`${plugin.resolve}/gatsby-node*`)[0]
 
 let apisRunningById = new Map()
 let apisRunningByTraceId = new Map()
@@ -157,29 +228,25 @@ module.exports = async (api, args = {}, pluginSource) =>
     })
 
     // Check that the API is documented.
-    if (!apiList[api]) {
+    // "FAKE_API_CALL" is used when code needs to trigger something
+    // to happen once the the API queue is empty. Ideally of course
+    // we'd have an API (returning a promise) for that. But this
+    // works nicely in the meantime.
+    if (!apiList[api] && api !== `FAKE_API_CALL`) {
       reporter.panic(`api: "${api}" is not a valid Gatsby api`)
     }
 
     const { store } = require(`../redux`)
     const plugins = store.getState().flattenedPlugins
-    // Get the list of plugins that implement gatsby-node
-    if (!filteredPlugins) {
-      filteredPlugins = plugins.filter(plugin => hasAPIFile(plugin))
-    }
 
-    // Break infinite loops.
-    // Sometimes a plugin will implement an API and call an
-    // action which will trigger the same API being called.
-    // "onCreatePage" is the only example right now.
-    // In these cases, we should avoid calling the originating plugin
-    // again.
-    let noSourcePluginPlugins = filteredPlugins
-    if (pluginSource) {
-      noSourcePluginPlugins = filteredPlugins.filter(
-        p => p.name !== pluginSource
-      )
-    }
+    // Get the list of plugins that implement this API.
+    // Also: Break infinite loops. Sometimes a plugin will implement an API and
+    // call an action which will trigger the same API being called.
+    // `onCreatePage` is the only example right now. In these cases, we should
+    // avoid calling the originating plugin again.
+    const implementingPlugins = plugins.filter(
+      plugin => plugin.nodeAPIs.includes(api) && plugin.name !== pluginSource
+    )
 
     const apiRunInstance = {
       api,
@@ -207,9 +274,13 @@ module.exports = async (api, args = {}, pluginSource) =>
     } else if (api === `onCreatePage`) {
       id = `${api}${apiRunInstance.startTime}${args.page.path}${args.traceId}`
     } else {
+      // When tracing is turned on, the `args` object will have a
+      // `parentSpan` field that can be quite large. So we omit it
+      // before calling stringify
+      const argsJson = JSON.stringify(_.omit(args, `parentSpan`))
       id = `${api}|${apiRunInstance.startTime}|${
         apiRunInstance.traceId
-      }|${JSON.stringify(args)}`
+      }|${argsJson}`
     }
     apiRunInstance.id = id
 
@@ -225,7 +296,26 @@ module.exports = async (api, args = {}, pluginSource) =>
       apisRunningByTraceId.set(apiRunInstance.traceId, 1)
     }
 
-    Promise.mapSeries(noSourcePluginPlugins, plugin => {
+    let stopQueuedApiRuns = false
+    let onAPIRunComplete = null
+    if (api === `onCreatePage`) {
+      const path = args.page.path
+      const actionHandler = action => {
+        if (action.payload.path === path) {
+          stopQueuedApiRuns = true
+        }
+      }
+      emitter.on(`DELETE_PAGE`, actionHandler)
+      onAPIRunComplete = () => {
+        emitter.off(`DELETE_PAGE`, actionHandler)
+      }
+    }
+
+    Promise.mapSeries(implementingPlugins, plugin => {
+      if (stopQueuedApiRuns) {
+        return null
+      }
+
       let pluginName =
         plugin.name === `default-site-plugin`
           ? `gatsby-node.js`
@@ -234,10 +324,16 @@ module.exports = async (api, args = {}, pluginSource) =>
       return new Promise(resolve => {
         resolve(runAPI(plugin, api, { ...args, parentSpan: apiSpan }))
       }).catch(err => {
+        decorateEvent(`BUILD_PANIC`, {
+          pluginName: `${plugin.name}@${plugin.version}`,
+        })
         reporter.panicOnBuild(`${pluginName} returned an error`, err)
         return null
       })
     }).then(results => {
+      if (onAPIRunComplete) {
+        onAPIRunComplete()
+      }
       // Remove runner instance
       apisRunningById.delete(apiRunInstance.id)
       const currentCount = apisRunningByTraceId.get(apiRunInstance.traceId)