+++ www/src/components/homepage/homepage-blog-post.js
@@ -1,28 +1,36 @@
 import React from "react"
 import PropTypes from "prop-types"
-import styled from "react-emotion"
+import styled from "@emotion/styled"
 import { Link, graphql } from "gatsby"
 import Img from "gatsby-image"
 
 import ArrowForwardIcon from "react-icons/lib/md/arrow-forward"
 
+import Avatar from "../avatar"
+
 import { HorizontalScrollerItem } from "../shared/horizontal-scroller"
 
-import presets, { colors } from "../../utils/presets"
-import { rhythm, options } from "../../utils/typography"
+import {
+  colors,
+  space,
+  radii,
+  transition,
+  shadows,
+  fontSizes,
+  breakpoints,
+} from "../../utils/presets"
+import { rhythm } from "../../utils/typography"
 
 const HomepageBlogPostRoot = styled(
   HorizontalScrollerItem.withComponent(`article`)
 )`
   display: flex;
   flex-direction: column;
-  font-family: ${options.systemFontFamily.join(`,`)};
   padding-bottom: ${rhythm(2.5)};
   position: relative;
 
-  .main-body & a {
+  a {
     border: none;
-    box-shadow: none;
     font-family: inherit;
 
     :hover {
@@ -30,41 +38,49 @@ const HomepageBlogPostRoot = styled(
     }
   }
 
-  ${presets.Tablet} {
-    width: 320px;
+  ${breakpoints.md} {
+    width: 20rem;
   }
 
-  ${presets.Desktop} {
+  ${breakpoints.lg} {
     flex-shrink: 0;
     margin-right: 0;
-    margin-bottom: ${rhythm(presets.gutters.default)};
+    margin-bottom: ${space[8]};
     padding-bottom: ${rhythm(3.5)};
     width: ${props => (props.fullWidth ? `100%` : `80%`)};
+    transition: transform ${transition.speed.default}
+        ${transition.curve.default},
+      box-shadow ${transition.speed.default} ${transition.curve.default};
 
     :hover {
-      background: ${colors.ui.whisper};
+      transform: translateY(-${space[1]});
+      box-shadow: ${shadows.overlay};
+    }
+
+    :active: {
+      box-shadow: ${shadows.cardActive};
+      transform: translateY(0);
     }
   }
 `
 
 const Cover = styled(Img)`
-  border-radius: ${presets.radiusLg}px ${presets.radiusLg}px 0 0;
+  border-radius: ${radii[2]}px ${radii[2]}px 0 0;
   display: block;
-  margin-bottom: -${rhythm(0.5)};
+  margin-bottom: -${space[3]};
 `
 
 const Header = styled(`h1`)`
   color: ${colors.gatsbyDarker};
-  font-size: 1.25rem;
+  font-size: ${fontSizes[4]};
   font-weight: bold;
-  line-height: 1.2;
   margin: 0;
   padding: ${rhythm(4 / 5)};
   padding-bottom: 0;
 
-  ${presets.Desktop} {
-    font-size: ${props => (props.first ? `1.75rem` : `1.5rem`)};
-    padding: ${rhythm(1.5)};
+  ${breakpoints.lg} {
+    font-size: ${props => (props.first ? fontSizes[6] : fontSizes[5])};
+    padding: ${space[7]};
     padding-bottom: 0;
   }
 `
@@ -74,17 +90,17 @@ const Meta = styled(`div`)`
   color: ${colors.gray.calm};
   display: flex;
   flex-wrap: wrap;
-  font-size: 0.875rem;
-  margin-top: 1rem;
+  font-size: ${fontSizes[1]};
+  margin-top: ${space[4]};
   padding: 0 ${rhythm(4 / 5)};
 
   & > * {
     flex-shrink: 0;
   }
 
-  ${presets.Desktop} {
-    margin-top: 1.5rem;
-    padding: 0 ${rhythm(1.5)};
+  ${breakpoints.lg} {
+    margin-top: ${space[6]};
+    padding: 0 ${space[7]};
   }
 `
 
@@ -93,24 +109,16 @@ const Author = styled(Link)`
   display: flex;
   z-index: 1;
 
-  img {
-    border-radius: 50%;
-    height: 28px;
-    width: 28px;
-  }
-
   span {
     color: ${colors.gatsby};
     border-bottom: 1px solid ${colors.ui.bright};
-    box-shadow: inset 0 -2px 0px 0px ${colors.ui.bright};
-    margin-left: 0.5rem;
   }
 
   a& {
     font-weight: normal;
   }
 
-  ${presets.Desktop} {
+  ${breakpoints.lg} {
     :hover {
       span {
         background: ${colors.ui.bright};
@@ -121,14 +129,12 @@ const Author = styled(Link)`
 
 const Excerpt = styled(`p`)`
   color: ${colors.gray.copy};
-  font-size: 0.875rem;
-  line-height: 1.5;
   padding: 0 ${rhythm(4 / 5)};
 
-  ${presets.Desktop} {
+  ${breakpoints.lg} {
     margin: 0;
-    margin-top: 1.5rem;
-    padding: 0 ${rhythm(1.5)};
+    margin-top: ${space[6]};
+    padding: 0 ${space[7]};
   }
 `
 
@@ -139,7 +145,7 @@ const ReadMore = styled(Link)`
   color: ${colors.gatsby};
   display: flex;
   flex-grow: 1;
-  font-size: 0.875rem;
+  font-size: ${fontSizes[1]};
   left: 0;
   padding: ${rhythm(4 / 5)};
   position: absolute;
@@ -152,20 +158,19 @@ const ReadMore = styled(Link)`
   }
 
   svg {
-    height: 18px;
-    width: 18px;
+    height: ${space[4]};
+    width: ${space[4]};
   }
 
   span {
     color: ${colors.gatsby};
     border-bottom: 1px solid ${colors.ui.bright};
-    box-shadow: inset 0 -2px 0px 0px ${colors.ui.bright};
     font-weight: bold;
-    margin-right: 0.2rem;
+    margin-right: ${space[1]};
   }
 
-  ${presets.Desktop} {
-    padding: ${rhythm(1.5)};
+  ${breakpoints.lg} {
+    padding: ${space[7]};
 
     span {
       :hover {
@@ -215,8 +220,9 @@ const HomepageBlogPost = ({
 
   return (
     <HomepageBlogPostRoot fullWidth={fullWidth}>
-      {desktopViewport &&
-        cover && <Cover fluid={cover.childImageSharp.fluid} />}
+      {desktopViewport && cover && (
+        <Cover fluid={cover.childImageSharp.fluid} />
+      )}
 
       <Link to={slug}>
         <Header first={first} withCover={cover}>
@@ -226,7 +232,7 @@ const HomepageBlogPost = ({
 
       <Meta>
         <Author to={authorSlug}>
-          <Img fixed={authorFixed} alt={authorName} />
+          <Avatar image={authorFixed} alt={authorName} />
           <span>{authorName}</span>
         </Author>
         &nbsp;on&nbsp;
@@ -250,44 +256,7 @@ HomepageBlogPost.propTypes = {
 
 export const homepageBlogPostFragment = graphql`
   fragment HomepageBlogPostData on MarkdownRemark {
-    excerpt
-    fields {
-      slug
-    }
-    frontmatter {
-      excerpt
-      title
-      date
-      author {
-        id
-        fields {
-          slug
-        }
-        avatar {
-          childImageSharp {
-            fixed(
-              width: 30
-              height: 30
-              quality: 80
-              traceSVG: {
-                turdSize: 10
-                background: "#f6f2f8"
-                color: "#e0d6eb"
-              }
-            ) {
-              ...GatsbyImageSharpFixed_tracedSVG
-            }
-          }
-        }
-      }
-      cover {
-        childImageSharp {
-          fluid(maxWidth: 700, quality: 80) {
-            ...GatsbyImageSharpFluid_withWebp
-          }
-        }
-      }
-    }
+    ...BlogPostPreview_item
   }
 `