+++ www/src/components/page-with-plugin-searchbar.js
@@ -1,26 +1,26 @@
 import React, { Fragment } from "react"
 import PluginSearchBar from "./plugin-searchbar-body"
 import { rhythm } from "../utils/typography"
-import presets, { colors } from "../utils/presets"
-import { scrollbarStyles } from "../utils/styles"
+import { colors, breakpoints, sizes } from "../utils/presets"
 
 const PageWithPluginSearchBar = ({ isPluginsIndex, location, children }) => (
   <Fragment>
-    <section
+    <nav
       css={{
         ...styles.sidebar,
         // mobile: hide PluginSearchBar when on gatsbyjs.org/packages/foo, aka package README page
-        display: `${!isPluginsIndex && `none`}`,
+        display: !isPluginsIndex ? `none` : false,
       }}
+      aria-label="Plugin navigation"
     >
       <PluginSearchBar location={location} />
-    </section>
+    </nav>
     <main
       id={`reach-skip-nav`}
       css={{
         ...styles.content,
         // mobile: hide README on gatsbyjs.org/plugins index page
-        display: `${isPluginsIndex && `none`}`,
+        display: isPluginsIndex ? `none` : false,
       }}
     >
       {children}
@@ -33,28 +33,27 @@ const widthLarge = rhythm(16)
 
 const styles = {
   sidebar: {
-    height: `calc(100vh - ${presets.headerHeight})`,
-    width: `100vw`,
+    height: `calc(100vh - ${sizes.headerHeight})`,
+    width: `100%`,
     zIndex: 1,
-    top: `calc(${presets.headerHeight} + ${presets.bannerHeight} - 1px)`,
-    ...scrollbarStyles,
-    [presets.Tablet]: {
+    top: `calc(${sizes.headerHeight} + ${sizes.bannerHeight} - 1px)`,
+    [breakpoints.md]: {
       display: `block`,
       width: widthDefault,
       position: `fixed`,
-      background: colors.ui.whisper,
-      borderRight: `1px solid ${colors.ui.light}`,
+      background: colors.white,
+      borderRight: `1px solid ${colors.gray.border}`,
     },
-    [presets.Desktop]: {
+    [breakpoints.lg]: {
       width: widthLarge,
     },
   },
   content: {
-    [presets.Tablet]: {
+    [breakpoints.md]: {
       display: `block`,
       paddingLeft: widthDefault,
     },
-    [presets.Desktop]: {
+    [breakpoints.lg]: {
       paddingLeft: widthLarge,
     },
   },