+++ www/src/pages/docs/ssr-apis.js
@@ -1,10 +1,10 @@
 import React from "react"
 import { graphql } from "gatsby"
 import { Helmet } from "react-helmet"
-import sortBy from "lodash/sortBy"
+import { sortBy } from "lodash-es"
 
-import Functions from "../../components/function-list"
-import { rhythm, scale } from "../../utils/typography"
+import APIReference from "../../components/api-reference"
+import { space } from "../../utils/presets"
 import Layout from "../../components/layout"
 import Container from "../../components/container"
 import { itemListDocs } from "../../utils/sidebar/item-list"
@@ -12,7 +12,9 @@ import { itemListDocs } from "../../utils/sidebar/item-list"
 class SSRAPIs extends React.Component {
   render() {
     const funcs = sortBy(
-      this.props.data.file.childrenDocumentationJs,
+      this.props.data.file.childrenDocumentationJs.filter(
+        doc => doc.kind !== `typedef`
+      ),
       func => func.name
     )
     return (
@@ -24,15 +26,15 @@ class SSRAPIs extends React.Component {
           <h1 id="gatsby-server-rendering-apis" css={{ marginTop: 0 }}>
             Gatsby Server Rendering APIs
           </h1>
-          <h2 css={{ marginBottom: rhythm(1 / 2) }}>Usage</h2>
-          <p css={{ marginBottom: rhythm(1) }}>
+          <h2 css={{ marginBottom: space[3] }}>Usage</h2>
+          <p css={{ marginBottom: space[5] }}>
             Implement any of these APIs by exporting them from a file named
             {` `}
             <code>gatsby-ssr.js</code> in the root of your project.
           </p>
           <hr />
-          <h2 css={{ marginBottom: rhythm(1 / 2) }}>APIs</h2>
-          <ul css={{ ...scale(-1 / 5) }}>
+          <h2 css={{ marginBottom: space[3] }}>APIs</h2>
+          <ul>
             {funcs.map((node, i) => (
               <li key={`function list ${node.name}`}>
                 <a href={`#${node.name}`}>{node.name}</a>
@@ -42,7 +44,7 @@ class SSRAPIs extends React.Component {
           <br />
           <hr />
           <h2>Reference</h2>
-          <Functions functions={funcs} />
+          <APIReference docs={funcs} showTopLevelSignatures={true} />
         </Container>
       </Layout>
     )
@@ -56,7 +58,7 @@ export const pageQuery = graphql`
     file(relativePath: { regex: "/api-ssr-docs.js/" }) {
       childrenDocumentationJs {
         name
-        ...FunctionList
+        ...DocumentationFragment
       }
     }
   }