+++ www/src/templates/template-docs-local-packages.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { graphql } from "gatsby"
-import _ from "lodash"
+import { pick } from "lodash-es"
 
 import PackageReadme from "../components/package-readme"
 import Unbird from "../components/unbird"
@@ -30,7 +30,7 @@ class DocsLocalPackagesTemplate extends React.Component {
     return (
       <>
         <PackageReadme
-          page={markdownRemark ? _.pick(markdownRemark, `parent`) : false}
+          page={markdownRemark ? pick(markdownRemark, `parent`) : false}
           packageName={
             markdownRemark
               ? markdownRemark.fields.title
@@ -70,7 +70,7 @@ class DocsLocalPackagesTemplate extends React.Component {
         />
         <Unbird
           dataSetId="5c1ac24b4a828a169b6c235c"
-          publicKey={process.env.UNBIRD_FEEDBACK_KEY_PLUGINLIB}
+          publicKey={process.env.GATSBY_FEEDBACK_KEY_PLUGINLIB}
           feedbackPrompt="Have feedback on the Plugin Library?"
         />
       </>