+++ packages/gatsby-transformer-toml/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.8...gatsby-transformer-toml@2.1.9) (2019-04-04)
+
+### Bug Fixes
+
+- **gatsby-transformer-toml:** Fix createContentDigest usage ([#13118](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/issues/13118)) ([506d7c3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/commit/506d7c3)), closes [#13112](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/issues/13112)
+
+## [2.1.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.7...gatsby-transformer-toml@2.1.8) (2019-04-02)
+
+**Note:** Version bump only for package gatsby-transformer-toml
+
+## [2.1.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.6...gatsby-transformer-toml@2.1.7) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-transformer-toml
+
+## [2.1.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.5...gatsby-transformer-toml@2.1.6) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-transformer-toml
+
+## [2.1.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.4...gatsby-transformer-toml@2.1.5) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-transformer-toml
+
 <a name="2.1.4"></a>
 
 ## [2.1.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-toml/compare/gatsby-transformer-toml@2.1.3...gatsby-transformer-toml@2.1.4) (2019-01-08)